+++ lib/pg_hero/standalone/error_html.ex
@@ -1,0 +1,7 @@
+defmodule PgHero.Standalone.ErrorHTML do
+  @moduledoc false
+
+  def render(template, _assigns) do
+    Phoenix.Controller.status_message_from_template(template)
+  end
+end