+++ lib/pg_hero/standalone/router.ex
@@ -1,0 +1,18 @@
+defmodule PgHero.Standalone.Router do
+  @moduledoc false
+  use Phoenix.Router
+  import PgHeroWeb.Router
+
+  pipeline :browser do
+    plug :accepts, ["html"]
+    plug :fetch_session
+    plug :fetch_flash
+    plug :protect_from_forgery
+    plug :put_secure_browser_headers
+  end
+
+  scope "/" do
+    pipe_through :browser
+    pghero "/"
+  end
+end